fix(sinogram): Store data as double instead of int
For histograming the integer was appropriate but when backprojecting
the results won't be whole numbers.

+++ inc/Sinogram.h
@@ -63,10 +63,10 @@ public:
 
     return sinogram;
   };
-  int data[NANGLES][NTANG] = {0};
+  double data[NANGLES][NTANG] = {0};
 
 private:
-  Sinogram(int *data) : data(data) {};
+  Sinogram(double *data) : data(data) {};
 
   /// @brief Check if indexes are in range or throw out_of_range
   void check_range(int angle_idx, int tang_idx) {
